+++ src/assets/styles/components/who-we-us.scss
@@ -1,71 +1,69 @@
 .who-we-us {
-	color: $white-color;
 	width: $total-parent;
 	display: flex;
+	justify-content: center;
+	align-items: center;
+	padding: 200px 0px 100px 0px;
+	background-color: #f9f9f9;
 
 	&__section {
 		height: 310px;
+		max-width: 350px;
+		margin: 0px 35px;
 
-		@media only screen and (max-width: 768px) {
-			flex-direction: row;
-			justify-content: center;
-			padding-left: 45px;
-			padding-right: 45px;
+		&__icon {
+			height: 100px;
+			margin-bottom: 38px;
 		}
 
-		&__inner {
-			justify-content: space-between;
-			align-items: center;
-			width: $width-50;
+		&__info {
+			color: $secondary-color;
 
-			@media only screen and (max-width: 768px) {
-				width: initial;
-				flex-direction: column-reverse;
+			&__title {
+				font-family: 'Akzidenz Grotesk - light';
+				font-size: 25px;
+				font-weight: bold;
+				font-stretch: normal;
+				font-style: normal;
+				line-height: normal;
+				letter-spacing: 1.5px;
+			}
+
+			&__descent {
+				margin-bottom: 30px;
+			}
+
+			&__descent,
+			&__description {
+				font-family: 'Lato', sans-serif;
+				font-size: 16px;
+				font-family: 'Lato';
+				letter-spacing: 1px;
+				padding-right: 20px;
+				line-height: 20px;
+			}
+		}
+	}
+
+	@media only screen and (max-width: 992px) {
+		flex-direction: column;
+		justify-content: center;
+		align-items: center;
+		padding: 40px 30px;
+
+		&__section {
+			height: auto;
+			margin: 30px 0;
+
+			&__icon {
+				height: 85px;
+				margin-bottom: 10px;
 			}
 
 			&__info {
-				height: $total-parent;
-				color: $secondary-color;
-				width: $width-50;
-
-				@media only screen and (max-width: 768px) {
-					width: 100%;
-				}
-
-				&__title {
-					font-family: 'Akzidenz Grotesk - bold condenced';
-					font-size: 25px;
-					font-weight: bold;
-					font-stretch: normal;
-					font-style: normal;
-					line-height: normal;
-					letter-spacing: 1.5px;
-					margin-bottom: 5px;
-				}
-
 				&__descent {
-					margin-bottom: 15px;
-				}
-
-				&__descent,
-				&__description {
-					font-family: 'Lato', sans-serif;
-				}
-			}
-
-			&__icon {
-				text-align: center;
-				width: $width-50;
-
-				@media only screen and (max-width: 768px) {
-					width: 100%;
-					text-align: initial;
-				}
-
-				img {
-					@media only screen and (max-width: 768px) {
-						width: 76px;
-					}
+					font-size: 14px;
+					margin-bottom: 20px;
 				}
 			}
 		}
