+++ src/components/home/who-we-us/constants.js
@@ -1,0 +1,27 @@
+import AUTOMATIZA from '../../../assets/images/home/automatiza.png';
+import GESTIONA from '../../../assets/images/home/gestiona.png';
+import COTIZA from '../../../assets/images/home/cotiza.png';
+
+export const DATA = [
+    {
+        icon: AUTOMATIZA,
+        descent: 'Fácil y rápido',
+        title: 'Automatiza',
+        description:
+            'Construye tu especificación con documentación e información técnica de las principales soluciones constructivas y productos.',
+    },
+    {
+        icon: GESTIONA,
+        descent: 'Edita tus EETT',
+        title: 'Gestiona',
+        description:
+            'Revisa, agrega, elimina y modifica. Mantén tu especificación técnica actualizada en un solo lugar de rápida forma ordenada.',
+    },
+    {
+        icon: COTIZA,
+        descent: 'Valoriza tu proyecto',
+        title: 'Cotiza',
+        description:
+            'Cotiza los productos utilizados en tu especificación con los principales proveedores y valoriza tu proyecto.',
+    },
+];
